refactor: move Contact component as page
